+++ changelog.md
@@ -9,7 +9,7 @@
     - `wayback-machine-secret-key`
 - remove Archive Backend: `archive-li`
 - only show the archive link icon on the user's bookmark page if the the `archiveUrl` exists
-- adjust dark-mode link visted color
+- adjust dark-mode link visited color and toggle
 - hide status of starred, to-read, and archive properties of bookmarks when not viewed by the bookmark owner
 
 ## v0.0.30 (2026-05-24)